Initial transition cell prototype implementation on only the z-axis.

+++ DarkCanvas/Assets/Scripts/ProceduralTerrain/Noise/NoiseMap3D.cs
@@ -8,7 +8,7 @@
         /// <summary>
         /// 3D array of values that represent the noise function.
         /// </summary>
-        public sbyte[,,] Values { get; init; }
+        public float[,,] Values { get; init; }
 
         /// <summary>
         /// Minimum height map value in the 3D array.
